+++ pers/config.py
